+++ src/main/java/com/example/newsfeed/auth/service/AuthService.java
@@ -1,12 +1,11 @@
-package com.example.newsfeed.user.service;
+package com.example.newsfeed.auth.service;
 
+import com.example.newsfeed.auth.dto.AuthRequestDto;
 import com.example.newsfeed.exception.CustomException;
 import com.example.newsfeed.exception.ExceptionType;
 import com.example.newsfeed.auth.config.PasswordEncoder;
 import com.example.newsfeed.user.domain.User;
-import com.example.newsfeed.user.dto.AuthRequestDto.LoginRequestDto;
-import com.example.newsfeed.user.dto.AuthRequestDto.RegisterRequestDto;
-import com.example.newsfeed.user.dto.AuthResponseDto.LoginResponseDto;
+import com.example.newsfeed.auth.dto.AuthResponseDto.LoginResponseDto;
 import com.example.newsfeed.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,7 +17,7 @@ public class AuthService {
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
 
-    public void register(RegisterRequestDto requestDto) {
+    public void register(AuthRequestDto.RegisterRequestDto requestDto) {
         validateNotDuplicatedEmail(requestDto.getEmail());
 
         String encodedPassword = passwordEncoder.encode(requestDto.getPassword());
@@ -31,7 +30,7 @@ public class AuthService {
         userRepository.save(user);
     }
 
-    public LoginResponseDto login(LoginRequestDto requestDto) {
+    public LoginResponseDto login(AuthRequestDto.LoginRequestDto requestDto) {
         User user = userRepository.findActiveUserByEmail(requestDto.getEmail())
                 .orElseThrow(() -> new CustomException(ExceptionType.INVALID_CREDENTIALS));
